Old FontManager: ignore fonts with no mappings
Ordinarily, this shouldn't happen but, if it does, we don't want
to pollute the cache with a blank charset for the font (as we'd
prefer to rescan the font next time around in case whatever caused
it to have no mappings got fixed in the interim).

Existing caches containing blank charsets for such fonts will
remain untouched -- this is all best effort, so forcing a rescan
by bumping the cache version is not worth doing here.